Better logic for map_empty
the previous algorith was such that, if at one point of the iteration
a certain key is found to be in the wrong place, then we short-circuit
and evaluate the rest of the negations
`map_empty?(...)`
which exists as `or ...` in the code.

But there is another case: when none of those keys are found, and it's just
that the recursive cases were resolved to false. In that case, the map
is _actually_ not empty, and we don't want to compute this recursive call.